fixed type uint -> size_t

+++ bar/common/cmdoptions.c
@@ -1212,7 +1212,7 @@ LOCAL void expandMacros(char *line, uint lineSize, const char *template, uint te
                      )
 #else /* not NDEBUG */
   bool __CmdOption_init(const char        *__fileName__,
-                        ulong             __lineNb__,
+                        size_t            __lineNb__,
                         CommandLineOption commandLineOptions[]
                        )
 #endif /* NDEBUG */
@@ -1350,7 +1350,7 @@ LOCAL void expandMacros(char *line, uint lineSize, const char *template, uint te
                      )
 #else /* not NDEBUG */
   void __CmdOption_done(const char        *__fileName__,
-                        ulong             __lineNb__,
+                        size_t            __lineNb__,
                         CommandLineOption commandLineOptions[]
                        )
 #endif /* NDEBUG */
